pyproject.toml
[project]
name = "rag-chatbot"
version = "0.1.0"
description = "マルチチャットボット対応汎用RAGシステム"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Your Organization", email = "contact@example.com"}
]

dependencies = [
    # Web API
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.6",
    # LangChain関連
    "langchain>=0.1.0",
    "langchain-community>=0.0.20",
    "langchain-chroma>=0.1.0",
    "langchain-ollama>=0.1.0",
    "langsmith>=0.1.0",
    # ベクトルストア（軽量版、PyTorch不要）
    "chromadb>=0.4.18",
    # 文書処理
    "PyPDF2>=3.0.1",
    "python-docx>=0.8.11",
    "markdown>=3.5.1",
    # ユーティリティ
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.1.0",
    "tqdm>=4.66.0",
    "aiohttp>=3.12.15",
    "psutil>=7.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project.urls]
homepage = "https://github.com/your-org/rag-chatbot"
repository = "https://github.com/your-org/rag-chatbot"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
